Funct send notifications

+++ ConferenceApp/Controllers/ChatController.cs
@@ -304,17 +304,20 @@ namespace ConferenceApp.Controllers
                     if (ModelState.IsValid)
                     {
                         var eventTags = new List<EventTag>();
-                    for (var i = 0; i < chat.AvailableTags.Count; i++)
-                    {
-                        if (chat.AvailableTags[i].IsChecked)
+                        if (chat.AvailableTags != null)
                         {
-                            var tag = await _context.Tags.FirstOrDefaultAsync(m => m.Id == chat.AvailableTags[i].TagId);
-                            var eventTag = new EventTag() {Event = chat, EventId = chat.Id, Tag = tag, TagId = tag.Id};
-                            _context.Add(eventTag);
-                            eventTags.Add(eventTag);
+                            for (var i = 0; i < chat.AvailableTags.Count; i++)
+                            {
+                                if (chat.AvailableTags[i].IsChecked)
+                                {
+                                    var tag = await _context.Tags.FirstOrDefaultAsync(m => m.Id == chat.AvailableTags[i].TagId);
+                                    var eventTag = new EventTag() {Event = chat, EventId = chat.Id, Tag = tag, TagId = tag.Id};
+                                    _context.Add(eventTag);
+                                    eventTags.Add(eventTag);
+                                }
+                            }
                         }
-                    }
-                    chat.EventTags = eventTags;
+                        chat.EventTags = eventTags;
                         _context.Add(chat);
                         await _context.SaveChangesAsync();
                         return RedirectToAction(nameof(Details), new { id = chat.Id.ToString() });
@@ -385,26 +388,29 @@ namespace ConferenceApp.Controllers
                 try
                 {
                     var eventTags = new List<EventTag>();
-                    for (var i = 0; i < chat.AvailableTags.Count; i++)
+                    if (chat.AvailableTags != null)
                     {
-                        var existingEventTag = await _context.EventTags.FirstOrDefaultAsync(x => x.TagId == chat.AvailableTags[i].TagId && x.EventId == chat.Id);
-                        if (chat.AvailableTags[i].IsChecked)
+                        for (var i = 0; i < chat.AvailableTags.Count; i++)
                         {
-                            // si el tag está checkeado y ya exitía este eventTag, no hacer nada, sino crearlo
-                            if (existingEventTag == null)
+                            var existingEventTag = await _context.EventTags.FirstOrDefaultAsync(x => x.TagId == chat.AvailableTags[i].TagId && x.EventId == chat.Id);
+                            if (chat.AvailableTags[i].IsChecked)
                             {
-                                var tag = await _context.Tags.FirstOrDefaultAsync(x => x.Id == chat.AvailableTags[i].TagId);
-                                var newEventTag = new EventTag() {Event = chat, EventId = chat.Id, Tag = tag, TagId = tag.Id};
-                                _context.Add(newEventTag);
-                                eventTags.Add(newEventTag);
+                                // si el tag está checkeado y ya exitía este eventTag, no hacer nada, sino crearlo
+                                if (existingEventTag == null)
+                                {
+                                    var tag = await _context.Tags.FirstOrDefaultAsync(x => x.Id == chat.AvailableTags[i].TagId);
+                                    var newEventTag = new EventTag() {Event = chat, EventId = chat.Id, Tag = tag, TagId = tag.Id};
+                                    _context.Add(newEventTag);
+                                    eventTags.Add(newEventTag);
+                                }
                             }
-                        }
-                        else
-                        {
-                            // si no está checkeado y ya exitía este eventTag, eliminarlo, sino no hacer nada
-                            if (existingEventTag != null)
+                            else
                             {
-                                _context.EventTags.Remove(existingEventTag);
+                                // si no está checkeado y ya exitía este eventTag, eliminarlo, sino no hacer nada
+                                if (existingEventTag != null)
+                                {
+                                    _context.EventTags.Remove(existingEventTag);
+                                }
                             }
                         }
                     }
